Refactor some code and move the interface to /api
