+++ src/main/java/de/fearnixx/jeak/service/controller/connection/HttpServer.java
@@ -39,7 +39,7 @@ public class HttpServer {
         this.objectMapper.setVisibility(PropertyAccessor.SETTER, JsonAutoDetect.Visibility.NONE);
         this.objectMapper.setVisibility(PropertyAccessor.GETTER, JsonAutoDetect.Visibility.NONE);
         this.objectMapper.setVisibility(PropertyAccessor.FIELD, JsonAutoDetect.Visibility.NONE);
-        init(8723);
+        init(port);
     }
 
     public void init(int port) {
